+++ DroneComponentsInventory/Models/MotorsComponent.cs
@@ -10,10 +10,10 @@ namespace DroneComponentsInventory.Models
         public string Manufacturer { get; set; } = null!;
         [Required]
         public string Model { get; set; } = null!;
-        public int? StatorSizeMm { get; set; }
+        public double? StatorSizeMm { get; set; }
         public int? Kv { get; set; }
         public string? MountPattern { get; set; }
-        public int? WeightG { get; set; }
+        public double? WeightG { get; set; }
         public int? RecommendedVoltageS { get; set; }
         public double? RecommendedPropInch { get; set; }
         public int? MaxThrustG { get; set; }
