+++ DroneComponentsInventory/Models/VideoTransmitterComponent.cs
@@ -13,10 +13,10 @@ namespace DroneComponentsInventory.Models
         public string? Type { get; set; }
         public double? MaxPowerMw { get; set; }
         public double? VoltageInputS { get; set; }
-        public int? MountPatternMm { get; set; }
+        public double? MountPatternMm { get; set; }
         public string? ControlProtocols { get; set; }
         public string? AntennaConnector { get; set; }
-        public int? WeightG { get; set; }
+        public double? WeightG { get; set; }
         public double? Price { get; set; }
     }
 }
